TOOL: renames batch/ to util/
