+++ libs/core/data-access/src/lib/batch.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 
 import { Batch, BatchSet, BatchUpdate } from './firestore/batch';
 import { FirestoreService } from './firestore/firestore.service';
-import { uniqueDiff } from './batch/unique-diff';
+import { uniqueDiff } from './util/unique-diff';
 
 @Injectable({
   providedIn: 'root',
